requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
beautifulsoup4>=4.12.0

youtube-transcript-api>=1.0.0
yt-dlp>=2024.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
